refactor - removed cameraRoll folder, passed to multimedia folder
