--- src/services/cameraRoll/cameraRollTypes.ts
@@ -1,5 +1,0 @@
-export type PhotoListPaginated = {
-  photoList: string[];
-  cursor?: string;
-  hasNextPage: boolean;
-};
