+++ src/services/index.ts
@@ -2,6 +2,5 @@ export * from './authCredentials';
 export * from './toast';
 export * from './storage';
 export * from './searchHistory/useSearchHistory';
-export * from './cameraRoll';
 export * from './permission';
 export * from './multimedia';
