+++ src/services/multimedia/index.ts
@@ -1,2 +1,4 @@
 export * from './multimediaType.ts';
 export * from './multimediaService.ts';
+
+export * from './useMultimediaGetPhotos.ts';
