+++ src/services/multimedia/multimediaService.ts
@@ -1,4 +1,19 @@
+import {CameraRoll} from '@react-native-camera-roll/camera-roll';
+
+import {PhotoListPaginated} from './multimediaType.ts';
 import {ImageForUpload} from './multimediaType.ts';
+
+async function getPhotos(cursor?: string): Promise<PhotoListPaginated> {
+  const photoPage = await CameraRoll.getPhotos({first: 10, after: cursor});
+
+  const photoList = photoPage.edges.map(edge => edge.node.image.uri);
+
+  return {
+    photoList,
+    cursor: photoPage.page_info.end_cursor,
+    hasNextPage: photoPage.page_info.has_next_page,
+  };
+}
 
 function prepareImageForUpload(imageUri: string): ImageForUpload {
   //TODO: implement
@@ -10,4 +25,4 @@ function prepareImageForUpload(imageUri: string): ImageForUpload {
   };
 }
 
-export const multimediaService = {prepareImageForUpload};
+export const multimediaService = {prepareImageForUpload, getPhotos};
