+++ src/services/multimedia/multimediaType.ts
@@ -3,3 +3,9 @@ export type ImageForUpload = {
   name: string;
   type: 'image/png' | 'image/jpg';
 };
+
+export type PhotoListPaginated = {
+  photoList: string[];
+  cursor?: string;
+  hasNextPage: boolean;
+};
