+++ src/services/multimedia/useMultimediaGetPhotos.ts
@@ -3,9 +3,9 @@ import {useEffect, useState} from 'react';
 import {QueryKeys} from '@infra';
 import {useInfiniteQuery} from '@tanstack/react-query';
 
-import {cameraRollService} from './cameraRollService.ts';
+import {multimediaService} from './multimediaService';
 
-export function useCameraRoll(
+export function useMultimediaGetPhotos(
   hasPermission: boolean,
   onInitialLoad?: (imageUri: string) => void,
 ) {
@@ -13,7 +13,7 @@ export function useCameraRoll(
 
   const query = useInfiniteQuery({
     queryKey: [QueryKeys.CameraRollList],
-    queryFn: ({pageParam}) => cameraRollService.getPhotos(pageParam),
+    queryFn: ({pageParam}) => multimediaService.getPhotos(pageParam),
     getNextPageParam: ({cursor}) => cursor,
     enabled: hasPermission,
   });
